rename and clean up
