+++ core/amber/src/main/python/core/models/tuple.py
@@ -18,7 +18,3 @@ Tuple.register(pandas.Series)
 @dataclass
 class InputExhausted:
     pass
-
-
-if __name__ == '__main__':
-    assert issubclass(pandas.Series, Tuple)
