+++ core/amber/src/main/python/core/runnables/network_receiver.py
@@ -12,11 +12,18 @@ from proto.edu.uci.ics.amber.engine.common import PythonControlMessage, PythonDa
 
 
 class NetworkReceiver(Runnable, Stoppable):
+    """
+    Receive and deserialize messages.
+    """
+
     def __init__(self, shared_queue: InternalQueue, host: str, port: int, schema_map: Dict[str, type]):
-        super().__init__()
         self._proxy_server = ProxyServer(host=host, port=port)
 
+        # register the data handler to deserialize data messages.
+        @logger.catch(reraise=True)
         def data_handler(command: bytes, table: Table):
+            # TODO: OPTIMIZE:
+            #   change the API to use pandas.DataFrame instead of pyarrow.Table
             data_header = PythonDataHeader().parse(command)
             if not data_header.end:
                 input_schema = table.schema
@@ -32,6 +39,8 @@ class NetworkReceiver(Runnable, Stoppable):
 
         self._proxy_server.register_data_handler(data_handler)
 
+        # register the control handler to deserialize control messages.
+        @logger.catch(reraise=True)
         def control_handler(message: bytes):
             python_control_message = PythonControlMessage().parse(message)
             shared_queue.put(ControlElement(tag=python_control_message.tag, payload=python_control_message.payload))
@@ -41,10 +50,12 @@ class NetworkReceiver(Runnable, Stoppable):
     def register_shutdown(self, shutdown: callable) -> None:
         self._proxy_server.register("shutdown", ProxyServer.ack(msg="Bye bye!")(shutdown))
 
-    @logger.catch
+    @logger.catch(reraise=True)
+    @overrides
     def run(self) -> None:
         self._proxy_server.serve()
 
+    @logger.catch(reraise=True)
     @overrides
     def stop(self):
         self._proxy_server.shutdown()
