+++ core/amber/src/main/python/core/runnables/test_network_sender.py
@@ -4,8 +4,8 @@ from time import sleep
 import pytest
 
 from core.models.internal_queue import InternalQueue
-from core.threads.network_receiver import NetworkReceiver
-from core.threads.network_sender import NetworkSender
+from core.runnables.network_receiver import NetworkReceiver
+from core.runnables.network_sender import NetworkSender
 
 
 class TestNetworkSender:
