+++ core/amber/src/main/python/core/util/runnable/runnable.py
@@ -1,9 +1,10 @@
 from abc import abstractmethod
+
 from typing_extensions import Protocol
 
 
 class Runnable(Protocol):
 
     @abstractmethod
-    def run(self):
+    def run(self) -> None:
         """run some logic"""
